+++ src/common/interfaces.ts
@@ -158,15 +158,20 @@ export interface Theme {
   };
 }
 
+// Use to extract the type of an Array's item type
+export type ArrayElement<ArrayType extends readonly unknown[]> =
+  ArrayType extends readonly (infer ElementType)[] ? ElementType : never;
+
 export interface DateValueType {
-  date: string;
+  id: number | string;
   value: number;
+  date: Date;
 }
 
 export interface LineGraphType {
   width: number;
   height: number;
-  data: Array<DateValueType>;
+  data: DateValueType[];
   yAxisUnit?: string;
   xAxisUnit?: string;
   startDateTimeString?: string;