+++ src/components/LineChart/LineChart.stories.tsx
@@ -21,23 +21,28 @@ TemperatureThroughoutDay.args = {
   xAxisUnit: "Messdatum",
   data: [
     {
-      date: "2021-04-09T04:13:55.000Z",
+      id: 13.9,
+      date: new Date("2021-04-09T04:13:55.000Z"),
       value: 13.9,
     },
     {
-      date: "2021-04-09T08:43:01.000Z",
+      id: 20.4,
+      date: new Date("2021-04-09T08:43:01.000Z"),
       value: 20.4,
     },
     {
-      date: "2021-04-09T13:21:39.000Z",
+      id: 18.6,
+      date: new Date("2021-04-09T13:21:39.000Z"),
       value: 18.6,
     },
     {
-      date: "2021-04-09T17:09:10.000Z",
+      id: 20.9,
+      date: new Date("2021-04-09T17:09:10.000Z"),
       value: 20.9,
     },
     {
-      date: "2021-04-09T20:43:59.000Z",
+      id: 22.6,
+      date: new Date("2021-04-09T20:43:59.000Z"),
       value: 22.6,
     },
   ],
@@ -51,27 +56,33 @@ HumidityOnceADay.args = {
   xAxisUnit: "Messdatum",
   data: [
     {
-      date: "2021-04-10T04:13:55.000Z",
+      id: 70,
+      date: new Date("2021-04-10T04:13:55.000Z"),
       value: 70,
     },
     {
-      date: "2021-04-11T04:13:55.000Z",
+      id: 55,
+      date: new Date("2021-04-11T04:13:55.000Z"),
       value: 55,
     },
     {
-      date: "2021-04-12T04:13:55.000Z",
+      id: 61,
+      date: new Date("2021-04-12T04:13:55.000Z"),
       value: 61,
     },
     {
-      date: "2021-04-13T04:13:55.000Z",
+      id: 47,
+      date: new Date("2021-04-13T04:13:55.000Z"),
       value: 47,
     },
     {
-      date: "2021-04-14T04:13:55.000Z",
+      id: 68,
+      date: new Date("2021-04-14T04:13:55.000Z"),
       value: 68,
     },
     {
-      date: "2021-04-15T04:13:55.000Z",
+      id: 51,
+      date: new Date("2021-04-15T04:13:55.000Z"),
       value: 51,
     },
   ],
@@ -85,27 +96,33 @@ PaxcounterThroughoutDay.args = {
   xAxisUnit: "Messdatum",
   data: [
     {
-      date: "2021-04-10T07:13:55.000Z",
+      id: 49,
+      date: new Date("2021-04-10T07:13:55.000Z"),
       value: 49,
     },
     {
-      date: "2021-04-11T08:13:55.000Z",
+      id: 189,
+      date: new Date("2021-04-11T08:13:55.000Z"),
       value: 189,
     },
     {
-      date: "2021-04-12T09:13:55.000Z",
+      id: 790,
+      date: new Date("2021-04-12T09:13:55.000Z"),
       value: 790,
     },
     {
-      date: "2021-04-13T17:13:55.000Z",
+      id: 100,
+      date: new Date("2021-04-13T17:13:55.000Z"),
       value: 100,
     },
     {
-      date: "2021-04-14T18:13:55.000Z",
+      id: 1002,
+      date: new Date("2021-04-14T18:13:55.000Z"),
       value: 1002,
     },
     {
-      date: "2021-04-15T19:13:55.000Z",
+      id: 306,
+      date: new Date("2021-04-15T19:13:55.000Z"),
       value: 306,
     },
   ],
@@ -117,27 +134,33 @@ ValuesWithoutSuppliedUnit.args = {
   height: 300,
   data: [
     {
-      date: "2021-04-10T04:13:55.000Z",
+      id: 70,
+      date: new Date("2021-04-10T04:13:55.000Z"),
       value: 70,
     },
     {
-      date: "2021-04-11T04:13:55.000Z",
+      id: 55,
+      date: new Date("2021-04-11T04:13:55.000Z"),
       value: 55,
     },
     {
-      date: "2021-04-12T04:13:55.000Z",
+      id: 61,
+      date: new Date("2021-04-12T04:13:55.000Z"),
       value: 61,
     },
     {
-      date: "2021-04-13T04:13:55.000Z",
+      id: 47,
+      date: new Date("2021-04-13T04:13:55.000Z"),
       value: 47,
     },
     {
-      date: "2021-04-14T04:13:55.000Z",
+      id: 68,
+      date: new Date("2021-04-14T04:13:55.000Z"),
       value: 68,
     },
     {
-      date: "2021-04-15T04:13:55.000Z",
+      id: 51,
+      date: new Date("2021-04-15T04:13:55.000Z"),
       value: 51,
     },
   ],