+++ src/components/LineChart/index.tsx
@@ -17,12 +17,12 @@ import "moment/locale/de";
 
 moment.locale("de-DE");
 
-const getX = (d: DateValueType): Date => {
-  return moment.parseZone(d.date).toDate();
-};
-const getY = (d: DateValueType): number => d.value;
+const getX = (d: { date: Date }): Date => d.date;
+const getY = (d: { value: number }): number => d.value;
 // eslint-disable-next-line @typescript-eslint/unbound-method
-const bisectDate = bisector<DateValueType, Date>(d => new Date(d.date)).left;
+const bisectDate = bisector<DateValueType, Date>(
+  (a, b) => b.getTime() - a.date.getTime()
+).left;
 
 const tooltipStyles = {
   ...defaultStyles,
@@ -95,22 +95,24 @@ export const LineChart = withTooltip<LineGraphType, DateValueType>(
           | React.MouseEvent<SVGRectElement>
       ) => {
         const { x } = localPoint(event) || { x: 0 };
-        const x0 = xScale.invert(x - padding.left);
-        const index = bisectDate(data, x0, 1);
-        const d0 = data[index - 1];
-        const d1 = data[index];
-        let d = d0;
-        if (d1 && getX(d1)) {
-          d =
-            x0.valueOf() - getX(d0).valueOf() >
-            getX(d1).valueOf() - x0.valueOf()
-              ? d1
-              : d0;
+        const dateInX = xScale.invert(x - padding.left);
+        const dateIndex = bisectDate(data, dateInX, 1);
+        const itemLeft = data[dateIndex - 1];
+        const itemRight = data[dateIndex];
+        let tooltipData = itemLeft;
+
+        if (itemRight && getX(itemRight)) {
+          tooltipData =
+            dateInX.valueOf() - getX(itemLeft).valueOf() <
+            getX(itemRight).valueOf() - dateInX.valueOf()
+              ? itemRight
+              : itemLeft;
         }
+
         showTooltip({
-          tooltipData: d,
-          tooltipLeft: xScale(getX(d)) + padding.left,
-          tooltipTop: yScale(getY(d)) + padding.top,
+          tooltipData,
+          tooltipLeft: xScale(getX(tooltipData)) + padding.left,
+          tooltipTop: yScale(getY(tooltipData)) + padding.top,
         });
       },
       [xScale, data, showTooltip, yScale]
@@ -224,7 +226,6 @@ export const LineChart = withTooltip<LineGraphType, DateValueType>(
         {tooltipData && (
           <div>
             <TooltipWithBounds
-              key={Math.random()}
               top={tooltipTop - 40}
               left={tooltipLeft}
               style={tooltipStyles}