+++ src/lib/dateUtil/dateUtil.test.ts
@@ -19,6 +19,9 @@ describe("createDateValueArray", () => {
         measurements: [20],
       },
     ]);
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    expect(dateValueArray[0].date instanceof Date).toBe(true);
     expect(dateValueArray[0].date.toISOString()).toBe(
       "2021-04-08T20:32:49.796Z"
     );