+++ src/lib/dateUtil/index.ts
@@ -1,25 +1,32 @@
 import moment from "moment";
 import "moment/locale/de";
 import { definitions } from "@technologiestiftung/stadtpuls-supabase-definitions";
+import { DateValueType } from "@common/interfaces";
 moment.locale("de-DE");
 
-interface DateValueType {
-  id: number;
-  value: number;
-  date: moment.Moment;
+export interface ExtendedDateValueType extends DateValueType {
+  formattedDay: string;
+  formattedTime: string;
+  dateISOString: string;
 }
 
 export const createDateValueArray = (
   input: Pick<definitions["records"], "id" | "recorded_at" | "measurements">[]
-): DateValueType[] => {
-  const dateValueArray = input.map(({ id, measurements, recorded_at }) => ({
-    id,
-    value: measurements && measurements.length >= 1 ? measurements[0] : 0,
-    date: moment.parseZone(recorded_at),
-  }));
+): ExtendedDateValueType[] => {
+  const dateValueArray = input.map(({ id, measurements, recorded_at }) => {
+    const dateMoment = moment.parseZone(recorded_at);
+    return {
+      id,
+      value: measurements && measurements.length >= 1 ? measurements[0] : 0,
+      date: dateMoment.toDate(),
+      dateISOString: dateMoment.toISOString(),
+      formattedDay: dateMoment.format("DD. MMM YYYY"),
+      formattedTime: dateMoment.format("HH:mm:ss"),
+    };
+  });
   const sortedDateValueArray = dateValueArray.sort((a, b) => {
-    return a.date.valueOf() - b.date.valueOf();
-  }) as DateValueType[];
+    return a.date.getTime() - b.date.getTime();
+  });
   return sortedDateValueArray;
 };